Field of view (unfinished)

+++ _迷信_/Assets/_/Stuff/CodeMonkey/Utils/UtilsClass.cs
@@ -339,7 +339,7 @@ namespace CodeMonkey.Utils {
         }
         
 
-        public static Vector3 GetVectorFromAngle(int angle) {
+        public static Vector3 GetVectorFromAngle(float angle) {
             // angle = 0 -> 360
             float angleRad = angle * (Mathf.PI/180f);
             return new Vector3(Mathf.Cos(angleRad), Mathf.Sin(angleRad));
